+++ tests/test_card_custom_card.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+
+import munchkin
+
+from nose.tools import *
+from munchkin.card import Card
+from args import Args
+
+args = Args()
+card_data = [
+    ['a', 'b', 'c', 'd', 'e', 'f', 'g'],
+    ['1', '2', '3', '4', '5', '6', '7'],
+    ['Q', 'w', 'e', 'r', 't', 'Y', 'u'],
+    ['9', '8', '7', '6', '5', '4', '3']
+]
+custom_card = Card(args)
+
+def setup():
+    custom_card.generate_custom_card(card_data)
+
+def test_custom_card_rows_and_columns():
+    eq_(custom_card.rows, 4)
+    eq_(custom_card.columns, 7)
+
+def test_custom_card_print_card():
+    ok_("| abcdefg |" in custom_card.m)
+    ok_("| 1234567 |" in custom_card.m)
+    ok_("| QwertYu |" in custom_card.m)
+    ok_("| 9876543 |" in custom_card.m)
+    ok_("|         |" in custom_card.m)
+    ok_("+-- 4x7 --+" in custom_card.m)